minor word changes
Calling them novice users rather than programmers to help distinguish this from the programming tutorials.  And I put a space between HPC and Carpentry.    There is also a "Supercomputing in Plain English" tutorial, but it goes into programming, MPI, OpenMP, etc pretty quickly, so we also want to make sure it's clear that this is different from that.